180 buildpower for carriers (= 3 air fac),
As for AA capacities, instead of thinking in terms of raw DPS compared to
Cruisers, It would be good if they bring something more, like longer range.
Carriers are supposed to be on the back of the fleet.

So I would suggest +25 range for all Carriers = 100 instead of 75,
which fit their secondary role of long range intel
(they have very wide vision radius and the equivalent of a t2 radar built in).
They are also slower than Cruisers, so it wouldn't be too easy to cover too
wide areas.
100 range is a lot but in line with the range of a t3 Battleship (128).

Additionally, I would suggest x2 or x2.5 DPS in the following manner:
- Aeon Keefer :
x2.5 damage of each shot = 300 (from 120) for a total of 600 DPS (from 240)
- Cybran Command Class =
x2 damage each shot = 20 (from 10) fro a total of 640 DPS (from 320)
- Seraphim Iavish :
damage = 60 (from 25) for a total of 480 DPS (from 200).
It should keep less AA DPS because it has a long range tac missile weapon.

As for Atlantis, it actually has very high AA DPS (1120),
but a relatively short range, so I would just suggest to make its AA range equal
 to 100 (from 60). An Atlantis that has surfaced is much more vulnerable,
 so extra AA range would help it more than additionnal DPS.

--HG--
branch : 3621