Sparky :
Target buffs:

1. The Sparky's jammer is one of it's strengths, but it's hard to make it count
in all the places it could. Jamming's main use at the moment is sucking up shots
from bigger units or T2 PD. The change proposed is relatively simply, provided
I understand the code correctly.

Code: Select all
SelectionPriority = 3,
 - remove this line of code. This should enable players to simply select
 Sparkies along with their normal units, allowing them to move in formation,
 or general with standard armies, as more of a mobile jammer. More on this in
 a bit.

2.
Code: Select all
 OnStartReclaim = function(self, target)
        TConstructionUnit.OnStartReclaim(self, target)
        self:SetAllWeaponsEnabled(false)
    end,

    OnStopReclaim = function(self, target)
        TConstructionUnit.OnStopReclaim(self, target)
        self:SetAllWeaponsEnabled( true)
    end,


Remove this. The Sparky is an effective reclaimer, and it's a combat engineer.
It should be able to fire whilst reclaiming.

3.
Code: Select all
   OnStartBuild = function(self, unitBeingBuilt, order)
        --Disable the gun while building something
        self:SetWeaponEnabledByLabel('Riotgun01', false)
        TConstructionUnit.OnStartBuild(self, unitBeingBuilt, order)
    end,

    OnStopBuild = function(self)
        --Re-enable the gun after done building
        self:SetWeaponEnabledByLabel('Riotgun01', true)
        TConstructionUnit.OnStopBuild(self)
    end,


Remove this. This is a stronger buff. This allows the Sparky to keep firing
whilst building. This is also in keeping with it's role as a T2 combat engy.
It is a bit weird that a unit that is all about having multiple applications
can't chew gum and walk at the same time.

3.
Code: Select all
Intel = {
        JamRadius = {
            Max = 12,
            Min = 12,
        },
        JammerBlips = 5,
        RadarRadius = 30,
        VisionRadius = 20,


I think these values could also withstand some experimenting. The goal here
is to accentuate both the counter-intel role Sparkies should be playing, as
well as their Combat Engineeringness.

--HG--
branch : 3621