Shields dont take spill over damage from other shields if they're already hurt by the source of the shield spilling. Because of this shields dont get extra hurt by spill over damage shields from other nearby shields.
For this feature each bubble shield now remembers the damage dealt to it. When there's spill over damage incoming first we check whether we already received this damage. If yes, ignore the spill over damage.
The approach to use a small database with damage information for each shield is not the best way but without access to the engine (to modify damagearea) I dont think there's a good way around this.

--HG--
branch : fix-scu-bubbleshield-overlap