Adding SupComDataPath.lua and init.lua files.
Thinking more about this, I really dislike this approach, however, for a couple
of reasons:

 - It's brittle (If we change faforever structure, relative links back to
   fa_path break)
 - There will be duplicated code in all of the game repositories.

I do still think we should have an init.lua file for each 'game', but let's do
it in a smart way, where we don't duplicate code and make a maintance hell for
ourselves.

I'll write up a proposal and some lua code soon to address this.

Closes #52 for now, I'll open a new issue when I have a proposal.

Conflicts:
	init.lua