Integrated NetLag into FAF
Setting net_Lag to slightly above the highest ping among the players in a
game. This will reduce command lag, especially for 1v1 games.

Refactored the code to use average ping instead of a snapshot. It's also
taking variable pings into account by adding a headroom of std-dev*2.5.

Logging (although reduced) is still activated, this should be turned off
when everything seem to be working as intended