Reduced wreck debris projectiles
Reduced the amount of wreck debris projectiles with ~50%.

Every time a unit dies it spawns alot of projetiles which are randomly shot
away from the wreck. This takes unnecessasry toll on the simulation,
especially when there are units dying constantly all over the field.

Most of the time people won't even see these but they still take up as
much CPU-resources.