Eliminate BitmapCombo2
BitmapCombo2 was a near-copy of BitmapCombo that seemed to exist soley to allow the coloured box to be stretched to fill the whole control.

Since the BitmapCombo class already special-cases the colour-selection case, let's migrate a simplified version of the necessary evil hack into it and throw away the huge amount of duplicated code.
In any case, there was only one changed function in BitmapCombo2, so if this solution is unacceptable we could definitely rephrase BitmapCombo2 to be a lot shorter (just replace the one function).

Still, this kills a ton of code and doesn't affect the UI. The only real downside is that using a BitmapCombo to create a colour selector of a different size will fail: you'll need to come up with a cleverer solution to this problem first.