Attempt at preventing connectivity screen
People exiting with ALT+F4 cause games to hang for a minute or so until a
timeout is reached and everyone boots them.

This is intensely annoying. The problem seems not to occur when a player
gracefully quits from the menu. This patch is a ridiculous hack that causes the
"graceful exit" code path to execute when a user simply ALT+F4s the game in
rage.

Note from Sheeo:
> All right, merging this since it ought to be harmless. We'll collect
> anecdotes during the lifetime of 3639 to see if it needs to go in 3640 or
> not.

Closes #248