Amusing hack to prevent quitting players from stopping games.
People exiting with ALT+F4 cause games to hang for a minute or so until a timeout is reached and everyone boots them.

This is intensely annoying. The problem seems not to occur when a player gracefully quits from the menu. This patch is a ridiculous hack that causes the "graceful exit" code path to execute when a user simply ALT+F4s the game in rage.