Show resources in map previews
This commit introduces a new class to represent map previews, with the capability to show resource points as well.
This allows us to remove a lot of duplicated logic for handling the large map, as well as adding resource previewing to the map previews in the map select screen and main lobby essentially for free.
To make the performance of the rendering of the resource icons acceptable, a mechanism for object-pooling sets of identical bitmaps is introduced (Avoiding the expensive operation of creating and destroying many Bitmaps whenever the map view tries to update, which would otherwise introduce considerable "jank" to the map selection screen).

This also makes the in-lobby map preview square.