Visual improvement of Lobby Options
Primarily I removed all the credits text, but also aligned the
checkboxes, capitalised, cleaned up the code a little bit.

Reasoning for credits removal: Xinnony is the main person behind this
lobby. Now that's it's open source and so many people are working on it,
there are too many to keep track of, so I think we should give Xinnony
and Barlots, his original partner, the only on-display credit, and keep
it in one place, in the primary lobby screen where it belongs.

Closes #273