Always generate cartographic view
Before this, if a map file doesn't contain cartographic colours, the preview
will not be generated.

Since the shaders are not being passed the water height level, the cartographic
preview cannot include the water level of the map, however.

There's a discussion thread on the forums about patching the engine for this
here: http://forums.faforever.com/forums/viewtopic.php?f=45&t=9141.