Prevent Flares drawing in allied entities
While I was dealing with the Depth Charges, decided to do the same with
flares. This will not affect much, but will stop Loyalist and Aeon TMD
redirecting allied missiles. It may do more, but the circumstances will
be very rare. Not sure if people want this 'quirk' of the Aeon TMD to be
changed, so this commit needs debate.