Minor autoKick logic cleanup
This was the other site that was calling UpdateGame before the UI was created. This was due to a call to SetCheck in the UI setup code for the checkbox invoking to OnCheck handler, which called UpdateGame.
The icing on the case is that UpdateGame never even *reads* autoKick, so it's an entirely redundant call anyway.

Finally, we make the "autoKick" flag stop being global.

Closes #379