Correctly populate blank profiles
Currently, if a host hosts a game using a profile that doesn't have sensible values for default options set on it, the options would be set to values equal to the index in the corresponding lobbyOptions value array where the default value lives.
This would prevent the game from starting, as the resulting values weren't valid settings. This would also prevent options from showing up in the box summarising game options.