A class for popup boxes
Instead of having an image for each popup box, let's use nine-patches to generate a background for arbitrarily-sized popups.
This requires the implementation of a nine-patch class.

GPG provides a class, Border, that gives us nine-patch-like behaviour, but it required identical images on opposite sides of the scaled region, and is quite deeply broken (it seems to always introduce gaps between the images, and occasionally odd scaling artefacts).
To make matters worse, the way GPG's code determines the size of Bitmaps for the purposes of layout appears defective, so the NinePatch implementation contains amusing fudge-factors to prevent gaps appearing between images...

"Elegant".