Fix UI corruption when watching replays
The "random" skin shall not be loaded in-game. This is enforced at game start (when the player who selected "random" is switched to the factional configuration that their randomness resolves to), and during skin rotation, but not when starting a replay, where it uses the skin for your last-used faction, which might be "random".

... So this ought to fix it. Probably.