Remove prefix sorting for map list
There was a mechanism that would group maps in the list by filename prefix.
This is somewhat expensive to calculate, and pretty unhelpful to actual users (maybe handy to developers who want to get their maps at the top).
In general, players expect maps to appear alphabetically, which they now do... Sort of.