Disable OnCollisionCheck instead of CollisionShape
This adresses hard-to-reproduce crash bugs with dying units.

My working theory is that projectiles attempting to do collision with units
that have recently had their CollisionShape set to 'None' will cause crashes
when OnCollisionCheck returns true, even though we are not meant to do
collisions.

Using DisallowCollisions on Unit has the same effect: Disallowing collisions,
but retains the collision shape.