Only TryCopyPose when death animation occurs
More testing of the pose errors reveals that it's due to some form of
garbage collection of wrecks. When there's been enough of unit deaths on
the field it starts to remove random wrecks.

This results in wrecks getting removed in the same tick as they are
created. This causes the error with TryCopyPose() when it tries to set a
pose of a deleted wreck.

It's unnecessary to copy pose of units that doesn't animated at death so
this fix at least reduce the amount of spam in log

I've not found out a way to raise the bar of the garbage collect. Maybe we
should implement our own GC of wrecks and delete the oldest ones instead,
this would prevent newly created wrecks to insta-disappear.