Added GC for wrecks
The in game garbage collection seem to delete wrecks by order of the
entity id. The problem is that the IDs seem to be recycled after a while,
so new wrecks get a low ID, resulting in them getting GCs before older
ones.

This patch overrides the default GC for wrecks. Lowest valued wrecks are
GC first and wrecks have a minimum lifetime of 2 minutes.