Have the autobalancer handle team "-" correctly
Team "-" represents "A special team containing only myself".
This is internally represented as the integer 1, not to be confused with team 1 (which is, naturally, represented using the integer 2).

The autobalancer was previously interpreting "-" (1) as being an ordinary team.
As a result, when you (because you're insane) asked the autobalancer to balance a free-for-all game, it would proceed to decide that everyone was on the same team and just randomly permute them.

Unfortunately, the autobalancer makes the assumption that the team numbering is sensible. Under older conditions, when asked to balance a two-team game, it would spit out a result in terms of team 1 and 2. This would lead to badness, because, naturally, team 1 is "No team". In abca674e82df248d854b99d7bcc97c348d28f6f2 we started incrementing the team output by the autobalancer to correct this.
Unfortunately, *this* means that when the autobalancer has decided, for a free-for-all, that everyone is on team 1, it then concludes that therefore everyone shall be asssigned to team 2 (that is, team 1), who then instantly win.

So, now, while building the input map of teams, we simply assign each person on team 1 ("-") an ascending integer for their team number, all to themselves. Testing shows the game doesn't seem to care about large team numbers.