Stop map border mesh from covering tall maps
It turns out that the map border comes as a collection of meshes that can only be scaled (in native code) preserving aspect ratio.
We have a square one, and one with width = 2 * height.
The scaling logic upscales the border according to the width of the map.

The previous logic would apply the wide one only when maps were exactly 2048*1024, and give everything else a square.
This would look odd on wide maps, and entirely break tall ones.
Wide before:
https://www.dropbox.com/s/18nyxhc7tci2byg/Screenshot%202015-04-21%2003.08.35.png?dl=0

After:
https://www.dropbox.com/s/6ehqejdtzhjg90f/Screenshot%202015-04-21%2003.02.11.png?dl=0

Tall maps no longer have any map border, irrespective of settings, because until someone makes a 1x2 mesh we can't render one without either a binary patch or having it cover half the map and ruin everything.

*sigh*