Set TargetPriority for weapons when loading BPs
Instead of setting the TargetPriority manually in each blueprint we set default
values when loading the blueprint. By looking at the weapon spec we can
decide what's a reasonable priority should be.

This will make it much easier to modify target priorities in the future
and keep them synced between all units.

The goal for this iteration is not to change the actual target priority of
units, just to move it into this new system