NetLag now sends a worstPing-table of every client
With this information you can give better information about who's lagging
a game. I've added a "Worst ping" to the connectivity screen (F11).

More UI improvements is needed (name of the worst guy) but it's a start.

Replaced the beat-function with a thread to get the pings to update even when
the game is stalled (by someone timing out)