New option: Show reclaim value of wreckage / props
If this is enabled, value of props / wreckage are shown if Ctrl+Shift is
pressed simultaneously.

The value will be the initial value of the reclaim and partial reclaim
won't change it.

Especially useful when playing new maps without knowing where the reclaim
is.

Reclaim with mass value below 20 isn't shown

Conflicts:
	lua/sim/Unit.lua