Prioritize assisting buildings over repair
When assisting a damaged unit the current engine behavior is to prioritize
repair, for instance repairing the ACU instead of helping it build a pd.

With this change, if you want to repair a unit over assisting a builder you
need to give explicit repair command.

There are known minor quirks with this fix so the engineers don't behave
exactly as if they assisted an undamaged builder (if not in build range they
walk near the building before starting to assist)