Prioritise targeting center of SCUs
Target bone lists are prioritised: attackers will run down the list until they find a bone they have line of sight to, and shoot at
it.
Making the head be the first choice greatly increases the chance of a miss by a projectile weapon. This change makes units fire at the
center of the unit, if they can, and call back to firing at the twiddly bits round the edges.