Stop wreckage hitboxes being underground
This doesn't do much very interesting, but is somehow relevant to the native-code magic that detects when a building is of the same
type as a wreck you're building it on top of.
If we do ever want wreckage to block fire or such, some tweaking of hitboxes for wrecks specifically will be needed, as many wrecks
have radically different apparent profiles to the unit they represent.