Flag wrecks to check for exploit
If a builder rebuilds something (decided by a call to GetRebuildBonus) we
check in SetRebuildProgress if the actual bonus never was triggered. This
means that there's a risk for a disappearing wreck.

We then search for potential wrecks on the map, iff one of these wrecks
are getting destroyed right after (same tick) we just recreate it

This scan for wrecks will happen very, very seldom, thus should not impact the
overall performance of the sim

Fixes #623