Prevent remote wreck deletion
This isn't perfect, but most likely stops this from being a useful exploit.
By using the call to GetRebuildBonus as a notification that the engine intends to delete a wreck, and by checking for the existence of
an appropriate wreck at the build site, we can spot the case of the engine wanting to delete a wreck but with no actual wreck existing
on the build site.
In this case, the engine is about to delete a wreck we want to keep. We then traverse all reclaimables in the simulation to find one
which is *most likely* the one the engine has just incorrectly deleted and recreate it. Known amusing edgecases:
- If another wreck of the same type is destroyed in the same tick as the one we want to undelete, that one may be undeleted instead.
- Engineers reclaiming the wreck that gets deleted/undeleted will stop, as their target has been destroyed.
- If the build template is moved to another wreck of the same type then the original target wreck is still deleted. This could in
principle be avoided by keeping track of all wrecks in a shiny datastructure, but it doesn't seem worth the cost.