Refactor reclaim code so it uses correct values
The latest refactoring caused wrong values for wrecks due to engine
blackbox behavior:

The engine assumes that GetReclaimCosts() returns the max reclaim value of
a prop. It then calculates the actual reclaim value by taking <max mass> *
<health ratio> * <fraction reclaimed>

This is why the 0.9 was needed in the UI to get the correct numbers (all
wrecks starts with 90% health).

Instead of keeping track of ReclaimMass we now just keep track of
ReclaimLeft, which then can be used if you need to calculate how much
reclaim is left of a wreck.  ReclaimLeft is updated when a wreck is
damaged or partly reclaimed.