No more invisible laser beams
Fixes a long-time bug where beams originating inside the fog-of-war wouldn't render.

This workaround operates by temporarily exposing the source of the beam during the initiation tick, and flushing any extra intel this
granted when the beam is deactivated.
The practical upshot of which is that a unit firing a beam is exposed for the duration of the beam fire, but no persistent intel is
leaked. This has apparently been deemed an acceptable tradeoff.

Fixes #29