Fixing up unit view pane
It's not possible to have both a shield and a tac pac (right?), so I
moved them on top of each other at the bottom middle. The kill count was
returned to the top middle. Fuel is now alone at the bottom right, and
no longer covers up the kill count.