Fix escape handling
This makes sure that we ignore any registered escape handlers iff the handler
is called with quit_game=True, which is the default for an unintercepted escape
key event.

It also uses table.getn(escapeHandlers) rather than keeping this as a
module-level local. Errors with traceback are emitted if an escapehandler is popped at an
invalid time, and the removal is cancelled (Previously invalid state would
occur).

TODO: - Refactor the commandmode reset + deselection code into a new keyaction (So it better fits the description of the 'escape' key)
- Take into account that users can remap the escape key to hit a different function than this. Or disallow remapping the escape key.

Fixes #786