Attempt at solving beam not switching off
Isnt the best way to do this. Uses the WatchForHoldFire to cut the beam if hold fire, or if stopped being in an "attack order" state. But i cant figure out how to differentiate the state between idle, and automatic attack (no command given). So i let the function cut the beam and restart it every 1 sec. This makes the beam start sound play every 1 second though, so I mute it if the time between beam stop and restart is 1 tick.