Set priority to targetted units
When the user is manually targeting units it add the blueprint of the
targetted units to each weapon.

The effect of this is that units will have a "memory" of what to shoot,
for instance when you choose to attack ACU with colossus and then move
with it. It will then continue to attack the ACU before attacking anything
else.

There is unlimited use for thiss, you can for instance order attack on t1
tanks with ACU, when you then start to move the ACU it will prioritize
those kind of units before shooting at something else.

It will take into account if you queue up attack targets, effectively
making you able to create priority for groups of units.