Refresh intel for destroyed structures
There's a bug where intel blips for destroyed / upgraded / given
structures don't get upgraded for enemy brains. This is problematic because you
cannot target given / upgraded stuff.

This is a workaround for this. When a structure dies it checks the enemy
brains if someone has a blip for it, if so it gets a brief vision on the
unit the get the intel.

Fixes #98