Merge RegenPercent with Regen
This will ensure that stacked regen values are always calculated
correctly. It also gets rid of the floor/ceil mechanic. Both
regenpercent and that mechanic are only used by the Sera Regen Field.
The floor/ceil was a disaster, as it over-rides all other buffs
stupidly. Having two types of regen buff meant immediate clashes.

This should fix 99.999% of the bugs with the regen buff system which
have annoyed us since 2007.