Fixed bug with hit points of builder
If a navy building sank during the same tick as it was finished the
builder had it's HP reduced to near zero.

It assumed a structure that triggered OnStopBeingBuilt + no-collision set
was a finished upgrade. In this case the no-collision flag was set due to
the building sinking.

Using an explicit IsUpgrade flag now instead to really make sure the finished
building was the result of an upgrade.

The curious can take a look at replay 4175596 to see the bug, happens for
MrZcion around min 8