Fix bugs with the alternative DamageArea
Due to DamageArea @ DamageArea.lua was defined local it actually returned
the real DamageArea from that import, so half of the testing was on the
original funcftion.

brain:GetUnitsAroundPoints() sucks, can only return Enemy / Ally /
Neutral, not all of them at the same time, so need to use GetUnitsInRect
instead.

GetReclaimablesInRect also returns reclaimable, alive units so we make
sure it's a prop

Damage() won't do damage if instigator == unit so have to call OnDamage
manually in this case