Make sure vision for upgraded blips arent permanent
If a blip for some reason survives over 5 ticks we manually delete it.
Old blips of upgraded units sometime survive for some unknown reason,
maybe the blip structure are reused by the engine or similar.

Only one clear thread is spawned regardless of how many blips are cleared
so should be effective.