add option to lay drop multiple lines of bombs at once
you can configure spread and number of lines from the blueprint, so
thats cool.

we spawn child projectiles and add velocity perpendicular to the motion
of the aircraft so they spread out.

however there appears to be a slight problem with the actual code that
does the spreading - the direction and magnitude of the displacement
vector appears to be correct, but for some reason it doesnt actually
land there. i suspect mystery lateral acceleration breaking all my
maths. or me being retarded. ah well.