fix uef scu aoe upgrade removal incorrect numbers
well i dunno who is responsible but this should be a crime imo. anyway
when you add 2 and remove -4 then add 2 again you get 0 aoe so if you
try to get the upgrade, remove, and get it again it no longer works.
great.

now thats fixed, and also NewMaxRadius was missing in here, lukily it
was called with (bp.NewMaxRadius or 25) so it wasnt so terrible, but
still