Fix LODCutoff for Seraphim ACU weapon trail
Optimizing a change in
https://github.com/FAForever/fa/commit/8a987da5fab49c61409a5b9b71b82abff32381e4

The LODCutoff for the weaponemitter of the Serapim ACU was set to 1200.
But the ACU Unit itself has a LODCutoff of 175.

This will lead to show the weapopntrail while full zoomed out without
showing the ACU unit.
So i reduced the visibilty of the weapon to the same zoom level as the
unit.

seraphim_chronotron_cannon_projectile_fxtrail_01_emit.bp
Changed from 1200 to 175 (was 100 in vanilla)

seraphim_chronotron_cannon_projectile_fxtrail_02_emit.bp
Changed from 1200 to 175 (was 100 in vanilla)

seraphim_chronotron_cannon_projectile_fxtrail_03_emit.bp
Changed from 1200 to 175 (was 100 in vanilla)