Adjust t4 sub height to clear sea floor
so this adds a slider which adds a cushioning effect on top of engine
side depth setting to have a larger offset above the seabed.

this doesnt affect the units "true" position however, only the position
of the bones. the ui,hitbox and meshextents remain unchanged.

so things like pathfinding are untouched.