Stop cloak FX flickering on all intel units. Fixes #1674
This happens in the no power state. Having no power cycles through
all of the intel types, not just the ones you should be able to use.
This accidentally triggers the cloak FX function. Checking that the
unit was meant to have cloak fixes it.

Note - This fix won't work with cloakfield units (The FX didn't anyway)
Will need to find a solution for BlackOps and other mods who use that
broken function...