Fix for Cloaking
line 2461 "if unit and not unit.Dead and not
unit:IsIntelEnabled('Cloak') then"
This will now set the CloakFXWatcherThread also on the source unit, also
prevents a cloakfield on a unit that already has his own cloak.
line 2479 "if self and not self.Dead and not
self:IsIntelEnabled('Cloak') then"
Dont disable cloak In case a unit has left a cloakfiled with its own
cloak activated.