Update loadingtips.lua
I have proposed some changes, I feel that the instructional language used previously drew too much attention to the advice-giver, if we want this advice to feel integrated it must try to avoid breaking the immersion of the game. Basically I want to replicate the tone and language of an in-game tutorial rather than an internet forum with advice for players. I have also tried to include the reason with the advice otherwise to avoid condescension.

Blimey.