counterintel disableing intel unintentionally
This commit fixes the following two issues:

The monkeylord loses its sonar capabilities in case of a powerstall when
the stealthfield is active. If it is deactivated, the sonar continues to
work. fixes #896

If the cybran acu, sacu or the uef sacu has a counter intel upgrade
(stealth, cloak, jamming) and the player powerstalls, the omni of the
unit will be deactivated.