add dummy weapon to sera sniper bot
fixes #1319

the main weapon (the first one in the weapons array) is disabled when it is used in sniper mode, thus it wont stop when it gets in range. This commit adds a new weapon that does no damage but stops the unit at the correct range due to being enabled all the time.