fix incomplete ships not being destroyed
naval units have a death animation which is only played when they are
more than 50% completed. In the DeathThread the degree of completion is
not checked again, just the existence of this animation. If the unit is
not sufficiently completed, it doesnt sink with its animation and thus
the SeaBedWatcher function doesnt terminate since the unit is not
sinking with the normal sinkerthread.

fixes #2076
fixes #1045