fix blacklist for coop maps
when the automated script generated the map blacklist it used the maps
on the test server. However, the map version on the test server severely
got out of sync from the map versions in the production enviroment. This
lead to several coop missions shown as outdated even though the mission
itself was the latest version.