change method to determine presence of AI markers
instead of searching for a land path node with a specific name, search for a marker that has other markers listed as "adjacent" markers. These markers are usually path nodes (not necessarily land pathnodes). Markers with only one entry for this property correspond to chains, so searching for the separator (space) in this property yields markers with multiple adjacent markers.