Implement Disruptable Teleportation
A teleporting acu/sacu will create a signature on the other end, and
enemy units can target it. If it sustains enough damage then the
teleportation is cancelled.

This way, its possible for your opponent to cancel the teleportation of
enemies to some locations and makes it a less risky business.