Remove reclaim preventing death explosions
#WhereWouldWeBeWithoutEQ
yeah turned out to be a quick fix this one xD
pretty much the title. when a unit dies it should die normally, and not
be able to prevent its death explosion by reclaiming it. now it dies
correctly. reclaiming without killing should result in no explosions.

NOTE: im gonna be away for about a month after this commit so wont
respond to all the bugs that are gonna be dug up, also this is poorly
tested so some qa boys want to get on the case there. but yeah lets just
hope for the best.