Updated hotfix for AI Replacement
Switching the conditions around as mentioned in the PR should fix the issue. There's no need for any validation to make sure the AIReplacement option exists, as the conditional statement takes care of that anyway. This code will check if `AIReplacementOn` is the value of `ScenarioInfo.Options.AIReplacement`, else the ACU of the player who has abandoned the game will just explode. This should fix the issue in ladder, but will require testing.