Fix Novax crashing other mods
changing the script like that lead to mods that hook
defaultprojectileweapon to change something in there, have ambiguous
class definitions which is very difficult to fix if you dont know what
you are doing, and it locked out a bunch of modding opportunities. to
prevent this, there is a script free way of doing the same thing, with a
blueprint value.

what this really does is highlight the need to refactor the weapon
class. this only works because the Main function inside idle state waits
for the recoil manipulators, which lasts until the beam is destroyed via
the thread. its a complete mess really. but thats how all beam units
work.

there should be a proper blueprint value to end beam on retarget or not,
and the recoil should not stall the beam end sequence.