Adjust teleportation location away from units
this stops you from teleporting right next to the paragon. not intended
as a complete fix by itself, it just means that you cant put your acu
right next to the enemy acu/paragon and die in friendly fire crazyness.

if finds the two most expensive things within a certain range and tries
to move away from them. it wont move if there is a large height
difference in that direction.