sera ACU upgrade reorganisation
move regen aura from the right to the left. Allowing the combo gun + regen aura.
It remove the combo T2 + regen aura, that was useful for navy. But nano can do the survivability job.
Moving RAS from back to right, in order to allow several choice on the right arm.