Blacklist old nomads version
this is rather annoying since people keep trying to use it, and its
completely messed up. This doesnt even show up in the mods list ingame!
its been hidden in the mod vault but whoever still has it might still
think its a great idea to try and use it.