Fix Kennel lag and animation
Issue was, if you build more than 30 kennels and upgrade them at the same time, it was freezing the game for 2-3 seconds when new drones spawns.
The lag came from a table.deepcopy function that was making a deepcopy of the drone data.
The deepcopy was not only making a copy of the unit, it also did a copy of platoondata, Brain and all other tables that where only present by a pointer.

Also fixed a minor issue with the door animation. They only opens if a real buildcommand was issued.
Selecting the drone and just let them start was not opening the doors.
Now doors stay open if the first drone is detached and will close if all drones return.