change oc dmg/e formula
exponential increase in e cost for dmg made sense with very high dmg cap but then we had to change it to a 15k dmg cap due to oc scus or smh so im changing the ratio to a flat 3/1, might be rebalanced to 3.5-4/1. 3/1 is almost the same in numbers as it is rn but easier to predict and much easier to edit. Min dmg changed to 1666 to keep the min e drain of 5k per oc.