SingleRangeOverlay cleanup
This code was added to fix issue #94 or at least to make it less harmful. But code logic is broken and doesn't work as intended.

Also idea is bad overall, cause if you have 'global' rings disabled and this code works as intended, then you can't see any rings for multiple selected units at all.

And it's 2020 after all, gpus are good enough to handle range rings in most cases + we did some optimisations in engine.