Add support for custom map Props
This requires a change to the init file. running this without the init
file should result in no changes - nothing breaks but no new features.

This adds support for custom props, and more to maps, and allows passing
arbitrary data to blueprints.lua, which is insanely useful and has been
a limitation for a long time.

Currently this is only used to load the custom content in the current
map, but mods and maps could do loads of stuff with it. Such as:
1. adding mod settings to turn off/on changes
2. adding map settings for optional features, and props.
3.mods can edit the logic by hooking modblueprints, and so can have
total control over which parts of themselves get loaded.

a note on security - this might sound like it creates a hole but it
doesnt let you break anything that couldnt be broken before:
game preferences is arbitrarily readable from UI already,
no writing functionality that wasnt there before has been added.
any non-deterministic change in blueprints.lua results in desyncs, just
like before
the only new abuse is stealth modding stuff into maps to create cheat
maps, which was possible before with prop stacking, or custom map
scripts. and this requires you to know what youre doing.