Fix attack orders for torp bombers
#3001

Happens when target is getting killed and torpedo bomber is out of range. This is a special case for naval units, because we don't call :Destroy() on them instantly and have to wait for sink thread.

In this case attacker has to be in weapon range in order to reset its attack order properly. And that's the problem for some torpedo bombers because of their firing pattern (t2 aeon torp vs sub, or t3 torp vs everything, t2 sera vs moving targets sometimes). So increasing maxRadius for 1 tick is the only option I guess