add topological sorting for mods
The existing implementation of mod sorting seems to have several issues:

* `ModComp` comparison does not check the before and after lists of the
second argument

* sorting relies on `table.keys` from `lua/system/utils.lua` and through
it on builtin `table.sort`, which assumes that an order between any pair
of elements can be established directly, which is not the case of mods
(for example if `mod2` depends on `mod1` and `mod3` depends on `mod2`,
`ModComp` looking at before/after lists of `mod1` and `mod3` will not be
able to establish a relationship between them correctly)

* there appears to be no functionality to log a warning if mods declare
an inconsistent ordering, contrary to the specification on top of the
`lua/MODS.LUA` file

This change introduces a new function `ModTopSort`, which implements a
simple topological sorting algorithm by a depth-first traversal of
dependencies. A warning is logged in case a cycle is encountered. The
old `ModComp` function is no longer used anywhere and is removed.
`GetActiveModsFiltered`, which previously relied on `ModComp`, retains
the same API, returning mods in an arbitrary order when before/after
lists are inconsistent, but should sort things more reliably otherwise.