increase performance of table.empty and math.round
I discovered that many functions in this file are are implemented in a very inefficient way. These two were just the easiest to fix, which is why I took a few seconds of my time to do so.

The new `table.empty` no longer needs to iterate over the whole table to know whether it's empty. It's therefore O(1) as opposed to O(n).
The new `math.round` no longer converts the value to a string and back again. It also supports using negative values to round to the nearest digit on the left side of the decimal point, as opposed to only allowing positive values to round to sub-decimal values.

Quick performance comparison of math.round
```
old round 5,000,000 iterations:	4.74 sec
new round 5,000,000 iterations:	0.628 sec
```