Increase wait on NukeAI to enable Aeon nuke launcher time to complete firing animation.
The default WaitSeconds time of 10 does not allow the Aeon Nuke launcher enough time to complete its animation to fire a missiles. This results in the default AI issuing a clear commands request and canceling the launch sequence. Increased to 12 seconds to allow time to launch missile. Currently the Aeon Nuke launcher will never fire for an AI that is using this platoon function. Have corrected for the default AI and the Sorian AI NukeAI platoon function..