Fix Commander not resuming build
CDROverCharge function is causing the build to fail.
The function stops the CRD platoon and resume the build with its own repair command.
But if we call the ProcessBuildCommand with "true" while running the overcharge function we will build a new building without resuming the old build.
So we need to stop the ProcessBuildCommand and let the overcharge function resume the build.